Update
Added bot prefix instead of hard coding it into the commands. This will allow us to change to prefix much easier should we ever choose to.

In addition I added command logging, so that anyone using a command, is logged. This goes for anyone who tried to use admin commands and fails.

The last change was minor and is a fail safe for the test bot. When we are working on the bot, it will no longer listen for commands. This prevents duplicate replies if a user should attempt using commands while we are working.