Update
The bot was getting messy, so I have split the commands into their own folder and files. This makes working on the bot much easier, and improves readability ten fold. This also makes passing arguments easier :smiley:!

This update also solves a small bug that we coded in which was commands replying as long as your message started with the command prefix and the command name. It would do so even if you continued to type after the command name. This is not intended or wanted, and this update solves that.